requirements.txt
flask
flasgger
python-xero
python-dotenv